+++ src/main/kotlin/lotto/domain/LottoGenerator.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 class LottoGenerator {
