+++ src/main/kotlin/lotto/domain/Value.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 data class Value(val money: Int) {
