+++ src/main/kotlin/lotto/views/InputView.kt
@@ -1,7 +1,6 @@
-package lotto
+package lotto.views
 
 class InputView {
-
     fun insertMoney(): String {
         println("구입금액을 입력해 주세요.")
         val input = readLine()
