+++ src/test/kotlin/lotto/domain/LottoStoreTest.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
